# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Economia Com História API",
        "description": "Documentação completa da API para a plataforma de gamificação educacional Economia Com História. Utiliza autenticação via Bearer Token (Sanctum).",
        "contact": {
            "email": "suporte@economiacomhistoria.ao"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor API principal"
        }
    ],
    "paths": {
...
        "/admin/access-levels": {
            "get": {
                "tags": [
                    "AccessLevels"
                ],
                "summary": "Listar todos os níveis de acesso (Apenas Admin)",
                "operationId": "adminIndexAccessLevels",
                "responses": {
                    "200": {
                        "description": "Níveis de acesso obtidos com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AccessLevel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "AccessLevels"
                ],
                "summary": "Criar um novo nível de acesso (Apenas Admin)",
                "operationId": "adminStoreAccessLevel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "gold"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Ouro"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Acesso super premium"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "🏆"
                                    },
                                    "color_bg": {
                                        "type": "string",
                                        "example": "#fef08a"
                                    },
                                    "color_text": {
                                        "type": "string",
                                        "example": "#854d0e"
                                    },
                                    "requires_approval": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "auto_grant": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nível de acesso criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access level created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AccessLevel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            }
        },
        "/admin/access-levels/{id}": {
            "get": {
                "tags": [
                    "AccessLevels"
                ],
                "summary": "Visualizar um nível de acesso específico (Apenas Admin)",
                "operationId": "adminShowAccessLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível de acesso",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nível de acesso obtido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AccessLevel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "404": {
                        "description": "Nível de acesso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AccessLevels"
                ],
                "summary": "Eliminar um nível de acesso (Apenas Admin)",
                "operationId": "adminDeleteAccessLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível de acesso",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nível de acesso eliminado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access level deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao eliminar"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "404": {
                        "description": "Nível de acesso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "AccessLevels"
                ],
                "summary": "Atualizar um nível de acesso existente (Apenas Admin)",
                "operationId": "adminUpdateAccessLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível de acesso",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ouro Premium"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Acesso a todos os conteúdos do site"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "🌟"
                                    },
                                    "color_bg": {
                                        "type": "string",
                                        "example": "#fef08a"
                                    },
                                    "color_text": {
                                        "type": "string",
                                        "example": "#854d0e"
                                    },
                                    "requires_approval": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "auto_grant": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nível de acesso atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access level updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AccessLevel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "404": {
                        "description": "Nível de acesso não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AccessLevel": {
                "title": "Access Level",
                "description": "Nível de acesso/privilégios de documentos/categorias",
                "required": [
                    "id",
                    "name",
                    "requires_approval",
                    "auto_grant"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "restricted"
                    },
                    "name": {
                        "type": "string",
                        "example": "Restrito"
                    },
                    "description": {
                        "type": "string",
                        "example": "Requer validação manual",
                        "nullable": true
                    },
                    "icon": {
                        "type": "string",
                        "example": "🔒",
                        "nullable": true
                    },
                    "color_bg": {
                        "type": "string",
                        "example": "#ffb3ba",
                        "nullable": true
                    },
                    "color_text": {
                        "type": "string",
                        "example": "#5c0011",
                        "nullable": true
                    },
                    "requires_approval": {
                        "type": "boolean",
                        "example": true
                    },
                    "auto_grant": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Token de autenticação Sanctum. Envie no cabeçalho: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "session_token": {
                "type": "apiKey",
                "description": "Token de sessão enviado no cabeçalho X-Session-Token",
                "name": "X-Session-Token",
                "in": "header"
            }
        }
    },
...
}